# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Simple AI Gateway",
        "description": "LLM API \u7f51\u5173 \u2014 \u517c\u5bb9 OpenAI API \u534f\u8bae\uff0c\u652f\u6301\u591a\u4f9b\u5e94\u5546\u8def\u7531\u3001\u7194\u65ad\u3001\u9650\u6d41\u3001\u914d\u989d\u3001\u52a8\u6001\u914d\u7f6e\u7ba1\u7406\u4e0e Webhook \u4e8b\u4ef6\u901a\u77e5\u3002\n\n## API \u5206\u7ec4\n- **\u4e1a\u52a1\u7aef\u70b9**: `/v1/chat/completions`, `/v1/models`\n- **\u8ba4\u8bc1**: `/auth/*`\n- **\u7ba1\u7406**: `/admin/*`\n- **\u5185\u90e8**: `/internal/*`\n- **\u7cfb\u7edf**: `/healthz`\n\n## \u8ba4\u8bc1\u65b9\u5f0f\nBearer token: JWT (access token) \u6216 API Key",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        },
        {
            "ApiKeyAuth": []
        }
    ],
...
    "paths": {
...
        "/admin/system/resilience": {
            "put": {
                "tags": [
                    "admin-system-config-controller"
                ],
                "operationId": "putSystemResilience",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResilienceConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResilienceConfig"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/system/pricing": {
            "put": {
                "tags": [
                    "admin-system-config-controller"
                ],
                "operationId": "putSystemPricing",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PricingConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PricingConfig"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/system/operational": {
            "put": {
                "tags": [
                    "admin-system-config-controller"
                ],
                "operationId": "putSystemOperational",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OperationalConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OperationalConfig"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/system/limit": {
            "put": {
                "tags": [
                    "admin-system-config-controller"
                ],
                "operationId": "putSystemLimit",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LimitConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/LimitConfig"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/routes/{id}": {
            "put": {
                "tags": [
                    "admin-route-controller"
                ],
                "operationId": "putRoute",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RouteConfig"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RouteConfig"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-route-controller"
                ],
                "operationId": "deleteRoute",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/admin/routes": {
            "get": {
                "tags": [
                    "admin-route-controller"
                ],
                "operationId": "routes",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoutesResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ResilienceConfig": {
                "type": "object",
                "properties": {
                    "maxAttempts": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "retryableFailureThreshold": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "failureWindow": {
                        "type": "object",
                        "properties": {
                            "seconds": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "zero": {
                                "type": "boolean"
                            },
                            "nano": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "negative": {
                                "type": "boolean"
                            },
                            "positive": {
                                "type": "boolean"
                            },
                            "units": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "durationEstimated": {
                                            "type": "boolean"
                                        },
                                        "timeBased": {
                                            "type": "boolean"
                                        },
                                        "dateBased": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "openDuration": {
                        "type": "object",
                        "properties": {
                            "seconds": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "zero": {
                                "type": "boolean"
                            },
                            "nano": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "negative": {
                                "type": "boolean"
                            },
                            "positive": {
                                "type": "boolean"
                            },
                            "units": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "durationEstimated": {
                                            "type": "boolean"
                                        },
                                        "timeBased": {
                                            "type": "boolean"
                                        },
                                        "dateBased": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "slowCallRateThreshold": {
                        "type": "number",
                        "format": "float"
                    },
                    "slowCallDurationThreshold": {
                        "type": "object",
                        "properties": {
                            "seconds": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "zero": {
                                "type": "boolean"
                            },
                            "nano": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "negative": {
                                "type": "boolean"
                            },
                            "positive": {
                                "type": "boolean"
                            },
                            "units": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "durationEstimated": {
                                            "type": "boolean"
                                        },
                                        "timeBased": {
                                            "type": "boolean"
                                        },
                                        "dateBased": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "slidingWindowType": {
                        "type": "string"
                    },
                    "slidingWindowSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "waitDurationInOpenState": {
                        "type": "object",
                        "properties": {
                            "seconds": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "zero": {
                                "type": "boolean"
                            },
                            "nano": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "negative": {
                                "type": "boolean"
                            },
                            "positive": {
                                "type": "boolean"
                            },
                            "units": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "durationEstimated": {
                                            "type": "boolean"
                                        },
                                        "timeBased": {
                                            "type": "boolean"
                                        },
                                        "dateBased": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "permittedNumberOfCallsInHalfOpenState": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "bulkheadMaxConcurrent": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "bulkheadMaxWaitDuration": {
                        "type": "object",
                        "properties": {
                            "seconds": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "zero": {
                                "type": "boolean"
                            },
                            "nano": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "negative": {
                                "type": "boolean"
                            },
                            "positive": {
                                "type": "boolean"
                            },
                            "units": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "durationEstimated": {
                                            "type": "boolean"
                                        },
                                        "timeBased": {
                                            "type": "boolean"
                                        },
                                        "dateBased": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "retryMaxAttempts": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "retryWaitDuration": {
                        "type": "object",
                        "properties": {
                            "seconds": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "zero": {
                                "type": "boolean"
                            },
                            "nano": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "negative": {
                                "type": "boolean"
                            },
                            "positive": {
                                "type": "boolean"
                            },
                            "units": {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "durationEstimated": {
                                            "type": "boolean"
                                        },
                                        "timeBased": {
                                            "type": "boolean"
                                        },
                                        "dateBased": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "retryExponentialBackoffMultiplier": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
...
            "PricingConfig": {
                "type": "object",
                "properties": {
                    "models": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/ModelPricing"
                        }
                    },
                    "default": {
                        "$ref": "#/components/schemas/ModelPricing"
                    }
                }
            },
...
            "OperationalConfig": {
                "type": "object",
                "properties": {
                    "maintenanceMode": {
                        "type": "boolean"
                    },
                    "emergencyRateLimit": {
                        "$ref": "#/components/schemas/EmergencyRateLimit"
                    },
                    "maintenanceWhitelist": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "RouteConfig": {
                "type": "object",
                "properties": {
                    "provider": {
                        "type": "string"
                    },
                    "upstreamModel": {
                        "type": "string"
                    },
                    "scene": {
                        "type": "string"
                    },
                    "fallbackRoutes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "weight": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "enabled": {
                        "type": "boolean"
                    }
                }
            },
...
            "RoutesResponse": {
                "type": "object",
                "properties": {
                    "generatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "routes": {
                        "type": "object",
                        "additionalProperties": {
                            "$ref": "#/components/schemas/RouteView"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT access token in Authorization header: Bearer <token>",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Gateway client API key in Authorization header",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}